# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticketopia API
  version: 1.5.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Read-only event discovery across all Ticketmaster categories, shared with the
    web UI. Event, artist, and venue IDs are provider-scoped (ticketmaster:SOURCE_ID).
    Category and genre IDs are Ticketmaster
    taxonomy IDs. Data is cached, not a live inventory or durable event history.
    Opt-in Google authentication provides MariaDB-backed profiles and private
    preferences, private durable event saves, independent event interest, and public event recommendations.
    Interest identity is private by default; counts include private choices but
    participant lists and profile activity include only explicitly public choices.
    Account APIs use a browser session or a user-scoped API token;
    discovery remains public. Account APIs return 503 when accounts are disabled.
servers:
  - url: /
security: []
paths:
...
  /api/v1/recommendations:
    get:
      operationId: listCommunityRecommendations
      summary: Browse public recommendations by city and category
      description: |
        Public endorsements, one per account/event, with optional plain-text reasons
        and allowlisted profiles. Multiple people can independently endorse the same
        event. Ordered by original recommended_at, then binary event ID and account
        ID descending; edits do not bump publication ordering. Filters use retained
        event snapshots, not provider requests. No private saves, interest identities,
        email or preferences appear. All responses are private, no-store, so withdrawal
        affects subsequent reads. Only the listed parameters, each once, are accepted.
      parameters:
        - {name: city, in: query, description: 'Case-insensitive exact city name, trimmed; no wildcard matching.', schema: {type: string, maxLength: 120}}
        - {name: country, in: query, description: Optional two-letter country code., schema: {type: string, pattern: '^[A-Za-z]{2}$'}}
        - {name: category_id, in: query, description: Event segment ID; all or omission removes this filter., schema: {type: string, pattern: '^[A-Za-z0-9_-]{1,64}$'}}
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/RecommendationCursor'}
      responses:
        '200':
          description: Recent public endorsements, or an empty array.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendationList'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
  /api/v1/me/event-recommendations:
    get:
      operationId: listOwnEventRecommendations
      summary: List your published recommendations
      description: Original publication ordering and last-known event snapshots. Recommendations are public even when your interest is private. Cursors are owner-bound. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/RecommendationCursor'}
      responses:
        '200':
          description: Your endorsements, or an empty array.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendationList'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
  /api/v1/me/event-recommendations/{event_id}:
    parameters:
      - {name: event_id, in: path, required: true, schema: {type: string, pattern: '^ticketmaster:[A-Za-z0-9_-]{1,128}$'}}
    get:
      operationId: getOwnEventRecommendation
      summary: Read your recommendation for an event
      description: No query parameters. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      responses:
        '200':
          description: Your public recommendation and retained event snapshot.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendation'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '404': {$ref: '#/components/responses/RecommendationNotFound'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
    put:
      operationId: setOwnEventRecommendation
      summary: Publish a public recommendation or replace its optional reason
      description: |
        Explicitly publishes your display name/profile and optional reason on the
        event, community collection and public profile. Does not change Save,
        Interested, attendance or posts. Omitted reason or an empty string clears
        the reason; null, repeated/unknown fields, client owner/metadata/visibility
        and query parameters are rejected. Reasons use valid UTF-8 plain text up
        to 500 Unicode characters, with surrounding whitespace trimmed. New
        identity/snapshot/recommendation commit atomically. Edits/retries preserve
        recommended_at; unchanged reasons also preserve updated_at. Existing edits
        need no provider call. Cookie writes require CSRF and same-origin checks;
        bearer writes do not use ambient cookies. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: {type: string, maxLength: 500, default: ''}
      responses:
        '201':
          description: Public recommendation created.
          headers:
            Location: {description: Own recommendation resource., schema: {type: string}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendation'}
        '200':
          description: Existing recommendation, with its optional reason replaced.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendation'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '404': {$ref: '#/components/responses/RecommendationNotFound'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
    delete:
      operationId: withdrawOwnEventRecommendation
      summary: Withdraw your recommendation without changing other activity
      description: Empty body and no query parameters. Owner-only, idempotent 204 even when absent. Retains shared metadata and other accounts' recommendations, saves and interest. Cookie writes require CSRF and same-origin checks. All responses are private, no-store.
      security: [{sessionCookie: []}, {bearerToken: []}]
      parameters:
        - {$ref: '#/components/parameters/CSRFToken'}
      responses:
        '204': {description: Recommendation withdrawn or already absent.}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/CSRFRejected'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
  /api/v1/events/{event_id}/recommendations:
    get:
      operationId: listEventRecommendations
      summary: Read public event recommendations and their count
      description: Count means explicit recommendations, not interest or Helpful. Unknown valid event IDs return zero and an empty array without provider calls. Ordered by original publication and binary IDs descending. All responses are private, no-store.
      parameters:
        - {name: event_id, in: path, required: true, schema: {type: string, pattern: '^ticketmaster:[A-Za-z0-9_-]{1,128}$'}}
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/RecommendationCursor'}
      responses:
        '200':
          description: Public recommendations and total explicit endorsements for this event.
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/EventRecommendationList'}
                  - {type: object, required: [count]}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
  /api/v1/users/{user_id}/event-recommendations:
    get:
      operationId: listPublicProfileRecommendations
      summary: Read an account's public recommendations
      description: Public endorsements independently of interest privacy. Only allowlisted public profile fields are included. Original publication ordering, retained snapshots, and public-account-bound cursors. All responses are private, no-store.
      parameters:
        - {name: user_id, in: path, required: true, schema: {$ref: '#/components/schemas/AccountID'}}
        - {$ref: '#/components/parameters/InterestLimit'}
        - {$ref: '#/components/parameters/RecommendationCursor'}
      responses:
        '200':
          description: This account's public recommendations, or an empty array.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EventRecommendationList'}
        '400': {$ref: '#/components/responses/AccountBadRequest'}
        '404':
          description: Account not found, including an invalid account ID.
          content:
            application/problem+json:
              schema: {$ref: '#/components/schemas/Problem'}
        '503': {$ref: '#/components/responses/RecommendationsUnavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-ticketopia_session
      description: 'HttpOnly, Secure, SameSite=Lax cookie issued after Google sign-in. Local HTTP development uses ticketopia_session instead.'
    bearerToken:
      type: http
      scheme: bearer
      description: Opaque, expiring user-scoped API credential created in an authenticated browser. Not a Google access token or ID token.
  parameters:
    RecommendationCursor:
      name: cursor
      in: query
      description: Opaque next_cursor bound to this collection, owner/public audience, event and normalized community filters.
      schema: {type: string, maxLength: 1024}
    InterestLimit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      description: Required for every cookie-authenticated write. Read csrf_token from GET /api/v1/me. Not required with an API bearer token. Supplied Origin/Referer must match AUTH_BASE_URL.
      schema: {type: string}
  responses:
    RecommendationNotFound:
      description: This owner has no recommendation, or the event cannot be resolved for first publication.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    RecommendationsUnavailable:
      description: Accounts/recommendations are disabled, storage is unavailable, or event data cannot be resolved. No fabricated count or in-memory fallback.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    AccountBadRequest:
      description: Unsupported query, invalid JSON/form fields or values, or a request body exceeding 16 KiB.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthenticated:
      description: Missing, expired, revoked, wrong-kind, or invalid credentials. Credential management requires a browser session.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    CSRFRejected:
      description: Invalid or missing CSRF token, or a cross-origin cookie write.
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    EventRecommendation:
      type: object
      additionalProperties: false
      required: [event, profile, reason, recommended_at, updated_at, meta]
      properties:
        event: {$ref: '#/components/schemas/Event'}
        profile: {$ref: '#/components/schemas/PublicProfile'}
        reason: {type: string, maxLength: 500, description: 'Optional reason as plain text, never trusted HTML.'}
        recommended_at: {type: string, format: date-time, description: Original publication time; edits and retries preserve ordering.}
        updated_at: {type: string, format: date-time, description: Last reason change; unchanged retries preserve this time.}
        meta: {$ref: '#/components/schemas/Freshness'}
    EventRecommendationList:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/EventRecommendation'}}
        next_cursor: {type: [string, 'null']}
        count: {type: integer, minimum: 0, description: 'Total endorsements for the event; present only on event-scoped reads, independently of pagination.'}
...
    AccountID:
      type: string
      pattern: '^[0-9a-f]{32}$'
    PublicProfile:
      type: object
      additionalProperties: false
      required: [id, display_name, bio]
      properties:
        id: {$ref: '#/components/schemas/AccountID'}
        display_name: {type: string, minLength: 1, maxLength: 80}
        bio: {type: string, maxLength: 500}
...
    ResourceID:
      type: string
      pattern: '^ticketmaster:[A-Za-z0-9_-]{1,128}$'
    Freshness:
      type: object
      required: [data_as_of, stale]
      properties:
        data_as_of: {type: string, format: date-time, description: 'When this data was successfully collected, not when this response was served.'}
        stale: {type: boolean, description: A refresh failed or was paused and retained data is being served.}
    Source:
      type: object
      required: [provider, id, url]
      properties:
        provider: {type: string, const: ticketmaster}
        id: {type: string}
        url: {type: string, description: 'Original source URL, or an empty string if unavailable.'}
    NamedID:
      type: object
      required: [id, name]
      properties:
        id: {type: string}
        name: {type: string}
...
    NullableInstant:
      type: [string, 'null']
      format: date-time
    EventStart:
      type: object
      required: [date_time, local_date, local_time, timezone, date_tba, date_tbd, time_tba, no_specific_time]
      properties:
        date_time: {$ref: '#/components/schemas/NullableInstant'}
        local_date: {type: [string, 'null'], description: Provider local date; TBA/TBD flags override any placeholder value.}
        local_time: {type: [string, 'null'], description: Provider local time; time_tba and no_specific_time override placeholder values.}
        timezone: {type: [string, 'null']}
        date_tba: {type: boolean}
        date_tbd: {type: boolean}
        time_tba: {type: boolean}
        no_specific_time: {type: boolean}
    PublicSale:
      type: object
      required: [start, end, start_tbd]
      properties:
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
        start_tbd: {type: boolean}
    Presale:
      type: object
      required: [name, description, url, start, end]
      properties:
        name: {type: string}
        description: {type: string}
        url: {type: string}
        start: {$ref: '#/components/schemas/NullableInstant'}
        end: {$ref: '#/components/schemas/NullableInstant'}
    PriceRange:
      type: object
      required: [type, currency, min, max, fees_included]
      properties:
        type: {type: string}
        currency: {type: string}
        min: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        max: {type: [string, 'null'], description: Decimal amount preserved without floating-point conversion.}
        fees_included: {type: [boolean, 'null'], description: Null when the source does not establish fee inclusion.}
    Image:
      type: object
      required: [url, width, height, ratio, fallback]
      properties:
        url: {type: string}
        width: {type: integer}
        height: {type: integer}
        ratio: {type: string}
        fallback: {type: boolean}
    Artist:
      type: object
      required: [id, name, source]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
    Place:
      type: object
      required: [name, address, city, state, country, country_code, postal_code, latitude, longitude]
      properties:
        name: {type: string}
        address: {type: string}
        city: {type: string}
        state: {type: string}
        country: {type: string}
        country_code: {type: string}
        postal_code: {type: string}
        latitude: {type: string}
        longitude: {type: string}
    Venue:
      allOf:
        - {$ref: '#/components/schemas/Place'}
        - type: object
          required: [id, source, timezone]
          properties:
            id: {$ref: '#/components/schemas/ResourceID'}
            source: {$ref: '#/components/schemas/Source'}
            timezone: {type: string}
    Classification:
      type: object
      required: [primary, segment, genre, subgenre]
      properties:
        primary: {type: boolean}
        segment: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        genre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
        subgenre: {anyOf: [{$ref: '#/components/schemas/NamedID'}, {type: 'null'}]}
    Event:
      type: object
      required: [id, name, source, start, status, public_sale, presales, price_ranges, images, artists, venues, place, classifications, info, please_note]
      properties:
        id: {$ref: '#/components/schemas/ResourceID'}
        name: {type: string}
        source: {$ref: '#/components/schemas/Source'}
        start: {$ref: '#/components/schemas/EventStart'}
        status: {type: string, description: 'Source status (onsale, offsale, canceled, postponed, rescheduled) or unknown. Not seat availability.'}
        public_sale: {$ref: '#/components/schemas/PublicSale'}
        presales: {type: array, items: {$ref: '#/components/schemas/Presale'}}
        price_ranges: {type: [array, 'null'], items: {$ref: '#/components/schemas/PriceRange'}, description: 'Null means no advertised prices, not free admission.'}
        images: {type: array, items: {$ref: '#/components/schemas/Image'}}
        artists: {type: array, items: {$ref: '#/components/schemas/Artist'}, description: 'Provider attractions, including performers or teams. The field name is retained for compatibility; events without attractions return an empty array.'}
        venues: {type: array, items: {$ref: '#/components/schemas/Venue'}}
        place: {anyOf: [{$ref: '#/components/schemas/Place'}, {type: 'null'}]}
        classifications: {type: array, items: {$ref: '#/components/schemas/Classification'}}
        info: {type: string}
        please_note: {type: string}
    Problem:
      type: object
      required: [type, title, status, code, detail]
      properties:
        type: {type: string, const: 'about:blank'}
        title: {type: string}
        status: {type: integer}
        code: {type: string, enum: [invalid_query, event_not_found, provider_unavailable, invalid_input, unauthenticated, csrf_failed, invalid_sign_in, account_not_found, rate_limited, credential_limit, accounts_disabled, accounts_unavailable]}
        detail: {type: string}
        fields: {type: object, additionalProperties: {type: string}}